Rearranging files and folders, also added some icons
