+++ public/Main.js
@@ -16,9 +16,8 @@ function createWindow() {
     }
 
     mainWin.loadURL(
-        isDev
-            ? 'http://localhost:3000'
-            : `file://${path.join(__dirname, '../build/index.html')}`,
+        // `file://${path.join(__dirname, '../build/index.html')}`
+        isDev ? 'http://localhost:3000' : `file://${path.join(__dirname, '../build/index.html')}`,
     )
 }
